+++ CHANGELOG.md
@@ -1,8 +1,10 @@
 # 0.9.0
 
+- Added: `Raycast` system param allows immediate raycasting into the world.
 - Removed the `Intersection` component. Intersection data can be found using `RaycastMesh::intersections()` and `RaycastSource::intersections()`.
-- Changed: `Ray3d::from_screenspace` start from the near plane
-- Fixed: Raycasts do not hit bottoms of un-rotated `RayCastMesh`es
+- Changed: `Ray3d::from_screenspace` start from the near plane.
+- Fixed: Raycasts do not hit bottoms of un-rotated `RayCastMesh`es.
+- Changed: `DefaultPluginState` renamed to `RaycastPluginState`.
 
 # 0.8.0